// src/app/interfaces/CSVFields.ts
export interface CSvFileds {
    DocTypeName: string;
    MinPages: number;
    MaxPages: number;
    HavingPageNumbers: string;
    PageNumberFormat: string;
    AddIfPrevTypeisSame: string;
    Priority: number;
    Keyword: string;
    PagePosition: string; // First/Last
    Percentageofsimilarity: string;
    KeywordThresholdpercetage: string;
    ZoneArea: string;
    Casesensitive: number; // 0-small,1- Caps
};